Add files via upload
Inheritance using array of objects method.
Here, the concept of java inheritance is implied by using constructors, objects and using array of objects.
Most of the code available on inheritance and not based on array of objects method and hence is not sufficient to cover the overall concepts related to inheritance method. This will also help you in familiarizing array of objects method which is very beneficial for working with java.